requirements.txt
streamlit
PyPDF2
faiss-cpu
numpy
python-dotenv
google-generativeai